requirements.txt
MouseInfo==0.1.3
mss==6.1.0
numpy==1.21.4
opencv-python==4.5.4.60
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
PyYAML==6.0
discord>=1.7.3